requirements.txt
fastapi
uvicorn
python-dotenv
requests
beautifulsoup4
jinja2
python-multipart

sentence-transformers
faiss-cpu
numpy
torch
torchvision
torchaudio

groq
